+++ js/image/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/image
 
+## 0.22.13
+
+### Dependency updates
+
+- @gradio/client@1.15.6
+- @gradio/statustracker@0.10.15
+- @gradio/upload@0.16.11
+
 ## 0.22.12
 
 ### Dependency updates