+++ js/label/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/label
 
+## 0.5.17
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.15
+
 ## 0.5.16
 
 ### Dependency updates