+++ js/slider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/slider
 
+## 0.6.14
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.15
+
 ## 0.6.13
 
 ### Dependency updates